+++ tools/scientific_tools.py
@@ -1,0 +1,74 @@
+class ScientificTools:
+    """A collection of mock tools for our AI Scientist agents."""
+
+    @staticmethod
+    def search_arxiv(query: str) -> str:
+        """
+        Simulates searching the arXiv database for a given query.
+        Returns a mock list of relevant papers.
+        """
+        print(f"\n MOCK TOOL: Searching ArXiv for '{query}'...")
+        return (
+            "Found Papers:\n"
+            "- 'The role of Graph Neural Networks in molecular design' (2023)\n"
+            "- 'Advancements in protein folding prediction using Transformers' (2022)\n"
+            "- 'A survey of generative models for drug discovery' (2024)"
+        )
+
+    @staticmethod
+    def execute_python_code(code: str) -> str:
+        """
+        Simulates executing Python code in a sandboxed environment.
+        Returns a mock result of the execution.
+        """
+        print(f"\n MOCK TOOL: Executing Python code in a sandbox...")
+        # print("--- CODE ---\n" + code + "\n------------")
+        # In a real implementation, NEVER use exec() on LLM-generated code
+        # without extreme sandboxing and security measures.
+        return "Execution successful. Data saved to 'results.csv'.\nColumns: 'molecule_id', 'binding_affinity', 'toxicity_score'"
+
+    @staticmethod
+    def analyze_data(file_path: str) -> str:
+        """
+        Simulates analyzing a data file.
+        Returns a mock summary of the analysis.
+        """
+        print(f"\n MOCK TOOL: Analyzing data from '{file_path}'...")
+        return (
+            "Data Analysis Summary:\n"
+            "- 25 out of 100 tested molecules show high binding affinity (>0.9).\n"
+            "- A positive correlation was found between molecular weight and toxicity.\n"
+            "- Molecule 'MOL-012' shows the most promise with high affinity and low toxicity."
+        )
+
+    @staticmethod
+    def write_latex_paper(analysis_summary: str, hypothesis: str) -> str:
+        """
+        Simulates writing a scientific paper in LaTeX format.
+        """
+        print("\n MOCK TOOL: Compiling analysis into a LaTeX paper draft...")
+        return f"""
+\\documentclass{{article}}
+\\title{{A Study on Novel Molecules for Drug Discovery}}
+\\author{{AI Scientist Framework}}
+\\begin{document}
+\\maketitle
+
+\\section*{{Abstract}}
+This paper investigates novel molecular structures based on the hypothesis that {hypothesis}. Our computational analysis reveals several promising candidates for further study.
+
+\\section{{Introduction}}
+The search for effective and non-toxic drugs is a significant challenge in modern medicine. This work explores...
+
+\\section{{Methodology}}
+We designed a computational experiment to generate and evaluate 100 novel molecules based on our initial hypothesis.
+
+\\section{{Results}}
+Our analysis of the experimental data reveals the following key findings:
+{analysis_summary}
+
+\\section{{Conclusion}}
+The results support our initial hypothesis. Specifically, molecule 'MOL-012' warrants further in-vitro testing.
+
+\\end{{document}}
+"""
